[chore/#41] components 폴더에 typo 폴더 생성
