+++ src/shared/components/index.ts
@@ -1,12 +1,16 @@
-export { Typo } from './Typo';
-export { Display } from './Display';
-export { Body } from './Body';
-export { Head } from './Head';
-export { Caption } from './Caption';
+export { Typo } from './typo/Typo';
+export { Display } from './typo/Display';
+export { Body } from './typo/Body';
+export { Head } from './typo/Head';
+export { Caption } from './typo/Caption';
 export { SectionPage } from './SectionPage';
 export { Header } from './Header';
 export type { HeaderProps } from './Header';
 export { BottomActionButton } from './BottomActionButton';
+export { Button } from './Button';
 export { TimePickerBottomSheet } from './TimePickerBottomSheet';
 export type { TimePickerValue } from './TimePickerBottomSheet';
 export { Icon } from './Icon';
+export { ModalContainer } from './modal/ModalContainer';
+export { LogoutModal } from './modal/LogoutModal';
+export { RevokeModal } from './modal/RevokeModal';
